refactor: move the cli module to the separate folder
