+++ ciphering-cli-tool/lib/cli/cli-errors.js
